Refactored io retrievers
